requirements.txt
# Core dependencies
python-dotenv==1.0.0
langchain>=0.3.21
langchain-core>=0.3.49
langchain-groq==0.0.1
langchain-community>=0.3.20
langchain-chroma==0.0.1
langgraph>=0.2.34

# Vector store and embeddings
chromadb==0.4.22
sentence-transformers>=3.2.0
langchain-huggingface==0.0.1

# Rich CLI interface
rich==13.7.0

# Async support
aiohttp>=3.10.10
asyncio>=3.4.3

# Web search (optional)
tavily-python>=0.5.0

# Utilities
typing-extensions>=4.12.2
pydantic==2.6.1

# Development
pytest>=8.3.3
black>=24.10.0
isort>=5.13.2
mypy>=1.13.0

# Added from the code block
fastapi==0.109.2
uvicorn==0.27.1